moved files to correct folder and added constants for activity names
