+++ src/Elsa.CustomActivities/Activities/Constants.cs
@@ -1,0 +1,8 @@
+﻿namespace Elsa.CustomActivities.Activities
+{
+    public class Constants
+    {
+        public const string CurrencyQuestion = "CurrencyQuestion";
+        public const string MultipleChoiceQuestion = "MultipleChoiceQuestion";
+    }
+}
