+++ src/Elsa.Server/Features/Currency/CurrencyQuestionController.cs
@@ -1,4 +1,5 @@
-﻿using MediatR;
+﻿using Elsa.Server.Features.Currency.SaveAndContinue;
+using MediatR;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Elsa.Server.Features.Currency
