+++ src/Elsa.Server/Features/Currency/SaveAndContinue/SaveAndContinueResponse.cs
@@ -1,4 +1,4 @@
-﻿namespace Elsa.Server.Features.Currency
+﻿namespace Elsa.Server.Features.Currency.SaveAndContinue
 {
     public class SaveAndContinueResponse
     {
