+++ src/He.PipelineAssessment.UI/Constants.cs
@@ -1,0 +1,8 @@
+﻿namespace He.PipelineAssessment.UI
+{
+    public class Constants
+    {
+        public const string CurrencyQuestion = "CurrencyQuestion";
+        public const string MultipleChoiceQuestion = "MultipleChoiceQuestion";
+    }
+}
